create general protvista component
